Reduce memory used to store inode names (#1305)
Each inode currently stores two separate strings for the key and the
name (always contained in the key string), resulting in redundant memory
usage. This change introduces a new `ValidKey` type which avoids the
duplication by only storing the key and the offset of the name for O(1)
retrieval.
`ValidKey` (and the related type `ValidName`) also enforce validation
for the name and the whole key at construction time, allowing calling
code to rely on the strings to be well-formed.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>