Update the changelog for mountpoint-s3 (#1319)
As a follow-up for this
[PR](https://github.com/awslabs/mountpoint-s3/pull/1304#pullrequestreview-2682411235)
we add a new entry in the `mountpoint-s3` change log.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

Change log was updated. No version change requred.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Evgeny (Zhenia) Dolgii <evdolgy@amazon.com>