Fix race condition in GetObject that could result in empty responses (#1334)
Address an issue in the `Stream` implementation for
`S3GetObjectResponse` that could immediately return `None` (i.e.
terminate the stream) when detecting that the meta request had
completed, before returning previously received parts. Reported in
#1331.

The fix changes the mechanism used to extract the response body parts
and the request completion from the meta request callbacks. Instead of
multiple independent channels, it introduces a single channel that
supports multiple `S3GetObjectEvent`s. The events in the new channel
match the order in which the callbacks are invoked, which is guaranteed
by the CRT. The events channel also includes the `Headers` event,
avoiding the need of a separate channel to await for the headers to be
returned.

When using Mountpoint, an occurrence of this issue would result in a
read request failing with an `Input/output error`, with a warning entry
in the logs containing this message:
```
mountpoint_s3_fs::fuse: read failed with errno 5: get request failed: get request terminated unexpectedly
``` 
Note however that we were not able to trigger the issue in our tests.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Bug fix entry and increase patch version.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>