Refactor internal S3 client methods to initiate meta-requests (#1337)
Re-organize the group of internal methods in the S3 client used to
initiate CRT meta-requests. Follow-up to the changes in #1334.

With this change, `S3CrtClientInner` provides 4 methods:
* `meta_request_with_callbacks`: the lowest-level method, which allows
more customization through callbacks. It is the basis for the other
methods and it is used directly by `get_object` and `put_object`.
* `meta_request_with_body_payload`: simpler variant that returns the
response body. Used by `list_object` and `get_object_attributes`.
* `meta_request_with_headers_payload`: variant returning the headers,
but no body. Used by `head_object` and `put_object_single`.
* `meta_request_without_payload`: simplest variant, only returns error.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No. Internal change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>