Create a package with custom extension (#1340)
Add a `--pkg-extensions` CLI flag which specifies which packages to
build and how to name them. Package type is inferred from top-level
component of the provided extension.

Also add a validation script for such packages.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Signed-off-by: Volodkin Vladislav <vladvolodkin@gmail.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>