Reduce memory usage for strings in inode metadata (#1346)
Reduce memory usage for strings included in inode metadata, like object
keys, etags, and inode names. Using a `Box<str>` instead of a `String`
ensures that no slack capacity is wasted and saves the `usize` field to
keep track of the buffer capacity.

### Does this change impact existing behavior?

No functional changes.

### Does this change need a changelog entry? Does it require a version
change?

Yes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>