`GetObjectResponse` returns part content as `Bytes` rather than `Box<[u8]>` (#1348)
Modify the `GetBodyPart` type streamed from `GetObjectResponse` so that
it exposes the part content as a `Bytes` type, rather than as a
`Box<[u8]>`. This is an API breaking change for `mountpoint-s3-client`,
which will require minor adjustments for users consuming the part
content. The switch to `Bytes` will enable the introduction of different
buffer allocation strategies in future releases.

### Does this change impact existing behavior?

No functional changes, but it is a minor API breaking change.

### Does this change need a changelog entry? Does it require a version
change?

Yes, it changes `mountpoint-s3-client` public API.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>