Cherry-pick GetObject cancellation fix (#1355) and prepare for client release 0.13.3 (#1356)
This change include the fix introduced in #1355 and increased the
`mountpoint-s3-client` crate version to 0.13.3 in preparation of
publishing to [crates.io](https://crates.io).

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Bug fix entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>