Make CRT memory limit configurable in Mountpoint's S3 client (#1363)
This is useful for benchmarking Mountpoint client with different memory
limits.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

Yes, for the client. Updated minor version of `mountpoint-s3-client`.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Sahitya Damera <sahityad@amazon.com>