Inline ObjectInfo to ReaddirEntry (#1366)
The type `object_client::ObjectInfo` is `#[non_exhaustive]` so it cannot
be constructed outside of the defining crate. In future we may want to
construct `ReaddirEntry`-s without using the client, but from data
stored on disk. As a workaround we inline ObjectInfo to ReaddirEntry.

### Does this change impact existing behavior?

No. This change is an implementation detail of the `readdir` module.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>