Update benchmark chart max datapoints from 20 to 30 (#1372)
Before this change, benchmark graphs
(https://awslabs.github.io/mountpoint-s3/dev/bench/) show up to 20 data
points where each data point represents a previous commit. One instance
we review this is in a weekly meeting, and we feel that more data points
would provide more contextual information of what changed as we could
have in excess of 20 commits over a period of one or two weeks.

This change updates the graphs to maintain 30 data points at the expense
of clarity.

### Does this change impact existing behavior?

No changes to Mountpoint or its crates. This will allow future benchmark
runs to maintain 30 data points in graphs.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>