Compile instance types into rust function  (#1369)
Follow up to #1368.

Changes the script to automatically generate a single rust match
expression (wrapped in a function) to get the instance throughput,
instead of parsing a `json` file. Additionally, now queries all regions.

This allows for better maintainability in the future, as we can see any
changes to the throughput numbers by looking at the diff in the
generated function.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>