Validate ServerSideEncryption on construction (#1373)
Minor change to validate the server-side encryption configuration
specified in the CLI arguments when the `ServerSideEncryption` instance
is built, rather than in a separate function.

### Does this change impact existing behavior?

No changes.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>