Add `fstab` CLI parser (#1362)
Introduce support for invoking Mountpoint with fstab style arguments:
```
./mount-s3 example-bucket /mnt/mountpoint -o rw,auto-unmount,allow-root
```

### Does this change impact existing behavior?

No breaking changes

### Does this change need a changelog entry? Does it require a version
change?

Yes, but not yet. This is still WIP

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>