Add CRT memory limit config to prefetcher and uploader benchmarks (#1379)
In some benchmarking, we want to experiment with adjusting the CRT's
memory limiter to observe the change in throughput performance.

This change introduces CLI flags to the benchmark scripts (examples)
that allows us to directly configure the CRT memory limiter.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No, impacts benchmarking scripts only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>