Implement creation of the internal manifest (#1377)
Implement creation of an SQLite database from an iterator of manifest
entries (in future reading from a file; now from RAM in tests). For more
context see
[branch](https://github.com/vladem/mountpoint-s3/pull/7/files) where
manifest is used to load metadata of the objects.

### Does this change impact existing behavior?

No, only used in tests.

### Does this change need a changelog entry? Does it require a version
change?

No, only used in tests.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>