Use the Runtime type in the prefetcher (#1382)
Modify the prefetcher to use the `Runtime` type (previously
`BoxRuntime`) instead of a generic parameter implementing `Spawn`.

This change simplifies the type signatures for many types used by the
Prefetcher, including `ObjectPartStream` and `DataCache`
implementations, in a similar way as already done for the Uploader.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Yes, for `mountpoint-s3-fs`.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>