Extract ClientConfig from CliArgs (#1380)
The new `ClientConfig` type captures all the configuration settings used
to initialize the S3 client. A `ClientConfig` instance can be built from
the relevant arguments in `CliArgs`, integrated with the settings
detected from `InstanceInfo`.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>