Use manifest in readdir and lookup operations (#1383)
Use metadata stored in an sqlite database instead of s3, when performing
lookup and readdir.

### Does this change impact existing behavior?

No, only used in tests.

### Does this change need a changelog entry? Does it require a version
change?

No, only used in tests.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>