Extract file system configuration from CliArgs (#1387)
Minor change to extract a `S3FilesystemConfig` from `CliArgs`. Part of
the effort to move `CliArgs` out of the `mountpoint-s3-fs` crate.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>