Remove Prefetch trait (#1388)
Simplify the type signature of `S3Filesystem` and related types by
removing the `Prefetch` trait and replacing it with a single
`Prefetcher` implementation, which has an `ObjectClient` generic
parameter.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>