Add `disk_data_cache.disk_usage_mib` metric (#1392)
Add a metric to record the amount of space used by cache as estimated by
MP internally. Relevant for
https://github.com/awslabs/mountpoint-s3/issues/1389.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>