Update benchmark script to always copy env vars (#1394)
This commit changes the way environment variables are populated when
running the benchmark script. There's no reason for us not to copy
environment variables from the script into the launched subprocesses,
and adding the functionality allows us to manipulate settings like the
logging level without making changes to `benchmark/benchmark.py`.

### Does this change impact existing behavior?

This change updates the way when using benchmark scripts, FIO and
Mountpoint are launched in `benchmark/` to copy over the existing
environment.

### Does this change need a changelog entry? Does it require a version
change?

No, benchmark change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>