Combine configuration for supported data cache types (#1395)
Introduce a new `DataCacheConfig` type to combine configuration for the
data cache in Mountpoint. The new type can be configured to enable a
local disk cache, a shared cache in S3 Express One Zone, or both.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>