Change type of init flags to u64
Recent upstream kernel releases have added more than 32 bits worth of
flags for the init message by adding 'flags2' fields on the message
types.  In order to prepare for this change, let's first change the
types of our existing init flags to u64.

This is an API break, but most users won't notice, unless they were
explicitly specifying the type of the flags somewhere.  None of the
examples are impacted.

This change is limited to the high-level API that we expose to users.
At the kernel ABI level, we currently just convert u32 to u64 and back.
That will change in the following commits.