Enable credentials caching with `--profile` flag (#1398)
Add a caching layer to the profile credentials provider, enabled by
`--profile` flag.

This change should provide a fix/mitigation for
https://github.com/awslabs/mountpoint-s3/issues/1358.

### Does this change impact existing behavior?

Yes, credentials will be cached for up to 15 minutes, when `--profile`
flag is used.

### Does this change need a changelog entry? Does it require a version
change?

Yes, added. Version `1.17.0` is the correct one for this change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>