Update crate versions and change logs for next crate publish (#1405)
This change ensures that all crate versions are up-to-date for
publishing new crate releases.
It also ensures the change logs are updated (with some minor
reordering), and fixes some comments related to crate versioning.

### Does this change impact existing behavior?

This is version updates and changelog updates only - no.

### Does this change need a changelog entry? Does it require a version
change?

This is a changelog update and version change!

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>