Update read-path layer benchmarks to consistently report throughput in Gib/s (#1397)
This updates the `prefetch_benchmark` and `download_crt` to report
throughput consistently with the `client_benchmark`. Note, the upload
path is untouched - notably, uploader benchmarks format is quite
different from these in reporting and still uses MiB/s.

### Does this change impact existing behavior?

This updates the output of the read-path benchmarks to be consistently
formatted. There's no way to switch back to the old format.

### Does this change need a changelog entry? Does it require a version
change?

No, benchmark change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>