Add example for new configuration options and manifest (#1403)
This adds a new example to Mountpoint, which showcases how MP can be
configured via API. In this example, we use the API to set configuration
options parsed from a json file.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>