Add test demonstrating MP behavior with ABAC IAM policies (#1415)
This adds new tests to demonstrate/document the behavior of Mountpoint
when trying to implement attribute-based access control (ABAC). The
purpose here is to simply demonstrate the behavior, so that we can
understand current state/options.

### Does this change impact existing behavior?

No, new test only.

### Does this change need a changelog entry? Does it require a version
change?

No, new test only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>