Update contributing to address updates of 0.x.y patch versions (#1406)
The guidance on how to update dependencies (and their dependents) was
unclear. This change updates the contributing guide.

### Does this change impact existing behavior?

Doc change only.

### Does this change need a changelog entry? Does it require a version
change?

No, doc change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>