Add errno check to FS mock S3 tests (#1424)
Simple update to the test to check error number. We check this in other
parts of the code, but this provides integration testing using the mock
S3 HTTP server tests.

Provides some basic coverage related to #1422.

### Does this change impact existing behavior?

No, test change only.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>