Add cache block serialization benchmark (#1426)
Introduce a criterion benchmark to measure the performance of reading a
single cache block. In order to establish a performance baseline, we
also include a separate benchmark for reading a file the size of a
block.

Updates `criterion` to `v0.6.0`.

The benchmark can be run with this command:

```
cargo bench --bench cache_serialization
```

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>