Add fstab to GitHub CI (#1419)
Add integration tests to Github CI

Tests passed here:
https://github.com/awslabs/mountpoint-s3/actions/runs/15135395318/job/42545885483


### Does this change impact existing behavior?
- fstab now forces runs in the background instead of the foreground
- When ran through fstab, Mountpoint now ignores the `nodev` and
`nosuid` options, as we default to these capabilities.

No breaking changes, as fstab is not released.

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>