Improve safety checks when reading disk cache blocks (#1427)
When reading data from the disk cache, we were not checking the declared
length of strings (such as the S3 key or ETag) and data, potentially
leading to allocations for the wrong size in case of corruption of the
cache block. While the corrupted block would still be detected later by
the integrity check and never returned to the user, the read could still
cause memory overflow in the worst case.

This change reworks the deserialization of a cache block from disk and
ensures that the length of strings is always within a fixed limit
(`10000`, using `bincode` configuration) and the data size is checked
against the cache block size (1 MiB).

In addition, we updated the `bincode` crate to the latest version
(`2.0.1`).

### Does this change impact existing behavior?

The change affect the on disk cache format, but it does not result in
any behavior change for the user.

### Does this change need a changelog entry? Does it require a version
change?

Changelog entry in the `fs` crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>