Logging errors in a structured format (#1416)
Implement an `FileErrorLogger` and use it to write errors occurring
during supported fuse calls in `S3FuseFilesystem`. This logging is not
enabled in the main executable, only in the `mount_from_config` example.

### Does this change impact existing behavior?

No, only used in the example.

### Does this change need a changelog entry? Does it require a version
change?

No, only used in the example.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Signed-off-by: Volodkin Vladislav <vladvolodkin@gmail.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>