Package fstab file (#1442)
Draft PR because I want to remove the fstab feature outside this PR

Adds `mount.mount-s3` symlink to our rpm and deb installers. This file
is placed in `/usr/sbin` in the host when installed.

### Does this change impact existing behavior?

Yes, a new `mount.mount-s3` file is added during installation.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Simon Beal <simobeal@amazon.com>