Update prefetcher wait_for_read_window_increment to drain queue (#1425)
When reviewing the prefetcher logic, the
`wait_for_read_window_increment` method call in the S3 part stream was
identified as a potential issue. The logic currently pulls only one
increment from the queue of read window increments when large amounts of
data are being fetched. Today, this is likely not to cause an issue as
the read increments are much larger than the size of the parts emitted
by `part_stream`. However, it would cause issues if there were changes
here in future that resulted in increments smaller than those parts.

This change updates the method to drain all available increments and
return the new value to the caller. This ensures that the backpressure
mechanism doesn't wait for each part before processing only one window
increment event. A new test is added to verify this behavior is
guaranteed.

### Does this change impact existing behavior?

There should be no impact, other than fixing logic that currently is
unlikely to introduce performance changes outside of very large parts
sizes.

### Does this change need a changelog entry? Does it require a version
change?

No, there is no known customer impact of the current issue.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>