Fstab tests: Ignore empty directory (#1443)
Currently, our fstab tests can fail if run in an environment where the
output dir does not exist.
This PR changes the `rm -r` call to a `rm -rf` to ignore cases where the
directory is empty.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>