Update prefetch and backpressure documentation, minor code changes for clarity (#1440)
This change should not change any functionality, and only modifies
document comments or rewrites code for clarity and to demonstrate
assumptions.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No behavior changes expected, no need for any changelog or version
change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>