Write documentation on Mountpoint with S3 on Outposts (#1452)
Adds some documentation on Mountpoint's support for S3 on Outposts.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No - this was already supported but we were missing docs.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Simon Beal <simobeal@amazon.com>