Ensure cache blocks are written atomically (#1433)
Address an issue with cache block reads failing while a concurrent write
is in progress, observed for example in #1389 (see log entries in
[comment](https://github.com/awslabs/mountpoint-s3/issues/1389#issuecomment-2861696762)).
This change modifies `put_block` to write to a temporary file first and
then rename to the expected cache block file name.

In addition, this PR also addresses concurrency issues in tracking block
usage data for eviction: updates to `UsageInfo` were not previously
synchronized correctly with the operations on disk and we could end up
recording a new block write when in fact the block had been concurrently
deleted. Now we lock `UsageInfo` while performing file system
operations.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Bug fix entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>