Add mock-mount-s3 to benchmark/ scripts (#1332)
This change allows us to run our benchmark scripts in `benchmark/` using
the `mock-mount-s3` binary, which presents a Mountpoint file system
backed by an in-memory mock S3 client.

This change itself incorporates quite a few changes (which may have been
better suited as separate commits). There are some changes to
accommodate configuration of part sizes in `mock-mount-s3`, removal of
throughput limits (which is useful for benchmarking!), and finally
adding the configuration options to the benchmarking scripts.

This change does include some hardcoded objects being added to
`mock-mount-s3` which can accomodate the benchmarking scripts. This
means that if the object keys change, the files will be created by FIO
and "uploaded" / populated in memory, which probably isn't what you
want.

### Does this change impact existing behavior?

No, there are no changes to main Mountpoint code.

### Does this change need a changelog entry? Does it require a version
change?

No, no behavior changes new or existing.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>